+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - albums: show groups to move/copy/export items
+- albums: hide grouped albums containing hidden items only
 
 ## <a id="v1.13.0"></a>[v1.13.0] - 2025-05-12